pyproject.toml
[tool.poetry]
name = "certificates-monitor"
version = "0.1.0"
description = "Sistema de Monitoramento de Certificados"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "2.3.3"
pymysql = "1.1.0"
sshtunnel = "0.4.0"
apscheduler = "3.10.1"
python-dotenv = "^1.2.1"
paramiko = "3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"